external_agents/weather_agent/requirements.txt
# Core A2A dependencies
a2a-sdk>=0.1.0

# HTTP client for API calls
httpx>=0.24.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Async support
asyncio>=3.4.3
